+++ src/library/DIPS.Mobile.UI/Resources/Icons/IconsExtension.cs
@@ -4,11 +4,7 @@ namespace DIPS.Mobile.UI.Resources.Icons;
 public class IconsExtension : IMarkupExtension<ImageSource>
 {
     public IconName IconName { get; set; }
-    public ImageSource ProvideValue(IServiceProvider serviceProvider)
-    {
-        var icon=  IconLookup.GetIcon(IconName);
-        return icon;
-    }
+    public ImageSource ProvideValue(IServiceProvider serviceProvider) => Icons.GetIcon(IconName);
 
     object IMarkupExtension.ProvideValue(IServiceProvider serviceProvider)
     {