+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [7.0.0]
+- ListItem: Renamed SubTitle to Subtitle
+- SkeletonView: Added SkeletonView as a new component. This is placed in the Loading namespace.
+- Components app: New Loading page with Skeleton loading.
+- Markup extensions: Added Margin and Padding markup extensions.
+
 ## [6.4.0] 
 - Resources was updated from DIPS.Mobile.DesignTokens